更新 采购统计  增加采购任务状态  增加付款记录上传附件  增加 阿里云私有上传

https://codeup.aliyun.com/60e3c522f00dfa2fa4cab323/backstage/merge_request/6294905

+++ app/src/main/java/cn/atsoft/dasheng/crm/entity/InvoiceBill.java
@@ -34,7 +34,7 @@ public class InvoiceBill implements Serializable {
      * 金额
      */
     @TableField("money")
-    private Integer money;
+    private Long money;
 
     /**
      * 附件
@@ -97,11 +97,11 @@ public class InvoiceBill implements Serializable {
         this.invoiceBillId = invoiceBillId;
     }
 
-    public Integer getMoney() {
+    public Long getMoney() {
         return money;
     }
 
-    public void setMoney(Integer money) {
+    public void setMoney(Long money) {
         this.money = money;
     }
